requirements.txt
boto==2.36.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
django-grappelli==2.6.3
django-storages==1.1.8
django-storages-s3upload==0.1.6
django-toolbelt==0.0.1
gunicorn==19.2.1
Pillow==2.7.0
psycopg2==2.6
python-magic==0.4.6
static3==0.5.1
